Move ConstraintResolution from BaseConstraint.h to ConstraintResolution.h
