+++ SofaKernel/modules/SofaCore/src/sofa/core/behavior/BaseConstraint.h
@@ -23,8 +23,7 @@
 #define SOFA_CORE_BEHAVIOR_BASECONSTRAINT_H
 
 #include <sofa/core/behavior/BaseConstraintSet.h>
-
-
+#include <sofa/core/fwd.h>
 
 namespace sofa
 {
@@ -34,66 +33,6 @@ namespace core
 
 namespace behavior
 {
-
-/**
- *  \brief Object computing a constraint resolution within a Gauss-Seidel algorithm
- */
-
-class SOFA_CORE_API ConstraintResolution
-{
-public:
-    ConstraintResolution(unsigned int nbLines, double tolerance = 0.0);
-
-    virtual ~ConstraintResolution() {}
-
-    /// The resolution object can do precomputation with the compliance matrix, and give an initial guess.
-    virtual void init(int /*line*/, double** /*w*/, double* /*force*/);
-
-    /// The resolution object can provide an initial guess
-    virtual void initForce(int /*line*/, double* /*force*/);
-
-    /// Resolution of the constraint for one Gauss-Seidel iteration
-    virtual void resolution(int line, double** w, double* d, double* force, double * dFree)
-    {
-        SOFA_UNUSED(line);
-        SOFA_UNUSED(w);
-        SOFA_UNUSED(d);
-        SOFA_UNUSED(force);
-        SOFA_UNUSED(dFree);
-        dmsg_error("ConstraintResolution")
-                << "resolution(int , double** , double* , double* , double * ) not implemented." ;
-    }
-
-    /// Called after Gauss-Seidel last iteration, in order to store last computed forces for the inital guess
-    virtual void store(int /*line*/, double* /*force*/, bool /*convergence*/) {}
-
-    inline void setNbLines(unsigned int nbLines)
-    {
-        m_nbLines = nbLines;
-    }
-
-    inline unsigned int getNbLines() const
-    {
-        return m_nbLines;
-    }
-
-    inline void setTolerance(double tolerance)
-    {
-        m_tolerance = tolerance;
-    }
-
-    inline double getTolerance() const
-    {
-        return m_tolerance;
-    }
-
-private:
-    /// Number of dof used by this particular constraint. To be modified in the object's constructor.
-    unsigned int m_nbLines;
-
-    /// Custom tolerance, used for the convergence of this particular constraint instead of the global tolerance
-    double m_tolerance;
-};
 
 /**
  *  \brief Component computing constraints within a simulated body.
